Initial support for reactive flows
Flows are subscribable, possibly asynchronous sequences of values. As the
mathematical dual of streams, flows push values to the user instead of the
user pulling elements from them. Like streams, flows support a rich set of
combinators for transforming and composing flows.

This commit bumps the project Java version to 8.